kiosk_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch>=2.2.0
torchaudio>=2.2.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
pydantic>=2.5.0
python-json-logger>=2.0.0
psutil>=5.9.0
requests>=2.31.0